Former-commit-id: e5bce53136bec0797b7f99bfcea03b2caf1cc085 [formerly 19cc1a02fcb66416e93149e7d08ae7f5d4dec36c] [formerly 5f640b198b0b0d10de5247c12487d1769bb2cbbd] [formerly e5bce53136bec0797b7f99bfcea03b2caf1cc085 [formerly 19cc1a02fcb66416e93149e7d08ae7f5d4dec36c] [formerly 5f640b198b0b0d10de5247c12487d1769bb2cbbd] [formerly 40261521ef7550484d4cde5de9987b2a3b4f2132 [formerly 5f640b198b0b0d10de5247c12487d1769bb2cbbd [formerly 76d37568c0de066d1b8eaa714863aa08b4d8f0f0]]]]
Former-commit-id: 40261521ef7550484d4cde5de9987b2a3b4f2132
Former-commit-id: c22fa47c66560fe8d003b7eb4573e022c5b0633a [formerly 06d81f0aa9a0b347451593d160710c937c7e170b] [formerly 8ccd74878f53b2b302e80feb93ed1cbf5c42f362 [formerly ab952b1acdd3a7aa5c039c6b8cb9856ca193e2ac]]
Former-commit-id: bda16046a2e8e53246c51469d2e3e20cf99f7dd8 [formerly 976cf25ebc44be25a29d1b95fbacdfce050ec391]
Former-commit-id: 29bc47e2c0246e6a05788ccb9f11ec6490f1ec49

+++ edexOsgi/com.raytheon.uf.edex.purgesrv/src/com/raytheon/uf/edex/purgesrv/PurgeJob.java
@@ -98,13 +98,21 @@ public class PurgeJob extends Thread {
         // Flag used to track if this job has failed
         boolean failed = false;
         startTime = System.currentTimeMillis();
-        PurgeLogger.logInfo("Purging expired data...", pluginName);
+        if (this.purgeType.equals(PURGE_JOB_TYPE.PURGE_ALL)) {
+            PurgeLogger.logInfo("Purging all data...", pluginName);
+        } else {
+            PurgeLogger.logInfo("Purging expired data...", pluginName);
+        }
         PluginDao dao = null;
 
         try {
             dao = PluginFactory.getInstance().getPluginDao(pluginName);
             if (dao.getDaoClass() != null) {
-                dao.purgeExpiredData();
+               if (this.purgeType.equals(PURGE_JOB_TYPE.PURGE_ALL)) {
+                    dao.purgeAllData();
+                } else {
+                    dao.purgeExpiredData();
+                }
 
                 PurgeLogger.logInfo("Data successfully Purged!", pluginName);
 
@@ -121,10 +129,10 @@ public class PurgeJob extends Thread {
                                         "Unable to purge data.  This plugin does not specify a record class and does not implement a custom purger.",
                                         pluginName);
                     } else {
-                        if (this.purgeType.equals(PURGE_JOB_TYPE.PURGE_EXPIRED)) {
-                            dao.purgeExpiredData();
-                        } else {
+                        if (this.purgeType.equals(PURGE_JOB_TYPE.PURGE_ALL)) {
                             dao.purgeAllData();
+                        } else {
+                            dao.purgeExpiredData();
                         }
 
                         PurgeLogger.logInfo("Data successfully Purged!",
